Fix C compiler determination in build-cxx-project
"clang++-" contains "g++-" so depending on the order in which the
map was iterated (since hash table iteration varies from run-to-run),
"clang++-" could unintentionally become "clangcc-".

Changes the regex to be anchored to the start of the string to prevent
this. A different solution may be needed if other compiler pairs are
ever introduced in which the section to be replaced does not occur
at the start.